pyproject.toml
[tool.poetry]
name = "it-help-service"
version = "0.1.0"
description = ""
authors = ["al-ov73 <klever1987@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.6"
psycopg2-binary = "^2.9.9"
djangorestframework = "^3.15.2"
pillow = "^10.3.0"
djangorestframework-simplejwt = "^5.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
